Extract native review row components
