+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1598,6 +1598,25 @@ final class ParityGateTests: XCTestCase {
         XCTAssertFalse(shellText.contains("QuillCodeToolCardView"), "Workspace shell should not own tool-card timeline placement.")
     }
 
+    func testNativeReviewPaneUsesFocusedRowFiles() throws {
+        let paneText = try Self.appSourceText(named: "QuillCodeReviewPaneView.swift")
+        let fileRowText = try Self.appSourceText(named: "QuillCodeReviewFileRowView.swift")
+        let lineRowText = try Self.appSourceText(named: "QuillCodeReviewLineRowView.swift")
+
+        XCTAssertTrue(paneText.contains("struct QuillCodeReviewPaneView"), "Review pane shell should remain a focused root view.")
+        XCTAssertTrue(paneText.contains("QuillCodeReviewFileRowView"), "Review pane shell should compose the focused file-row view.")
+        XCTAssertTrue(fileRowText.contains("struct QuillCodeReviewFileRowView"), "Review file rows should live in a focused file.")
+        XCTAssertTrue(fileRowText.contains("QuillCodeReviewHunkView"), "Review hunk rows should live beside file-row rendering.")
+        XCTAssertTrue(fileRowText.contains("QuillCodeReviewActionButton"), "Review action buttons should live beside file and hunk row rendering.")
+        XCTAssertTrue(fileRowText.contains("QuillCodeReviewLineRowView"), "Review hunk rendering should compose the focused line-row view.")
+        XCTAssertTrue(lineRowText.contains("struct QuillCodeReviewLineRowView"), "Review line rows should live in a focused file.")
+        XCTAssertTrue(lineRowText.contains("markerColor"), "Line marker styling should live beside line-row rendering.")
+        XCTAssertTrue(lineRowText.contains("lineBackground"), "Line background styling should live beside line-row rendering.")
+        XCTAssertFalse(paneText.contains("struct QuillCodeReviewHunkView"), "Review pane shell should not own hunk rendering.")
+        XCTAssertFalse(paneText.contains("struct QuillCodeReviewLineRowView"), "Review pane shell should not own line rendering.")
+        XCTAssertFalse(paneText.contains("struct QuillCodeReviewActionButton"), "Review pane shell should not own row action controls.")
+    }
+
     func testWorkspaceSwiftUIViewDelegatesSheetPresentation() throws {
         let shellText = try Self.appSourceText(named: "WorkspaceSwiftUIView.swift")
         let sheetsText = try Self.appSourceText(named: "QuillCodeWorkspaceSheets.swift")
